REFACTOR - Big refactor with all Cosmos classes. start implementing User in CosmosDB
