+++ noq-backend/src/main/java/com/noq/backend/DTO/cosmos/UserCosmosDTO.java
@@ -1,0 +1,6 @@
+package com.noq.backend.DTO.cosmos;
+
+import com.noq.backend.models.cosmos.ReservationCosmos;
+
+public record UserCosmosDTO(String id, String name, ReservationCosmos reservation) {
+}
