+++ noq-backend/src/main/java/com/noq/backend/controllers/cosmos/BedController.java
@@ -1,6 +1,6 @@
 package com.noq.backend.controllers.cosmos;
 
-import com.noq.backend.DTO.BedDTO;
+import com.noq.backend.DTO.cosmos.BedDTO;
 import com.noq.backend.DTO.CreateBedRequest;
 import com.noq.backend.services.cosmos.BedService;
 import lombok.RequiredArgsConstructor;
