+++ noq-backend/src/main/java/com/noq/backend/controllers/cosmos/UserCosmosController.java
@@ -1,6 +1,8 @@
 package com.noq.backend.controllers.cosmos;
 
-import com.noq.backend.DTO.HostCosmosDTO;
+import com.noq.backend.DTO.cosmos.HostCosmosDTO;
+import com.noq.backend.DTO.cosmos.UserCosmosDTO;
+import com.noq.backend.models.cosmos.UserCosmos;
 import com.noq.backend.services.cosmos.UserCosmosService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -20,27 +22,27 @@ public class UserCosmosController {
     private final UserCosmosService userCosmosService;
 
 
-    // CREATE NEW HOST
+    // CREATE NEW USER
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public Mono<ResponseEntity<HostCosmosDTO>> create(@RequestBody HostCosmosDTO request) {
-        return hostCosmosService.create(request)
+    public Mono<ResponseEntity<UserCosmosDTO>> create(@RequestBody UserCosmosDTO request) {
+        return userCosmosService.create(request)
                 .map(response -> ResponseEntity.status(HttpStatus.CREATED).body(response))
                 .onErrorResume(error -> INTERNAL_SERVER_ERROR_RESPONSE);
     }
 
-    //GET HOST WITH ID
+    //GET USER WITH ID
     @GetMapping(path = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     public Mono<ResponseEntity<HostCosmosDTO>> getUserById(@PathVariable("id") String id,
-                                                           @RequestBody HostCosmosDTO request) {
-        return hostCosmosService.findById(request.hostId(), request.name())
+                                                           @RequestBody UserCosmosDTO request) {
+        return userCosmosService.findById(request.id(), request.name())
                 .map(response -> ResponseEntity.status(HttpStatus.OK).body(response))
                 .switchIfEmpty(NOT_FOUND_RESPONSE)
                 .onErrorResume(error -> INTERNAL_SERVER_ERROR_RESPONSE);
     }
 
-    //GET ALL HOSTS
+    //GET ALL USERS
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public Flux<HostCosmosDTO> getAllUsers() {
+    public Flux<UserCosmosDTO> getAllUsers() {
         return hostCosmosService.findAll().onErrorResume(error -> Mono.error(
                 new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Database not responding.")
         ));
