+++ noq-backend/src/main/java/com/noq/backend/services/cosmos/HostCosmosService.java
@@ -1,7 +1,7 @@
 package com.noq.backend.services.cosmos;
 
 import com.azure.cosmos.models.PartitionKey;
-import com.noq.backend.DTO.HostCosmosDTO;
+import com.noq.backend.DTO.cosmos.HostCosmosDTO;
 import com.noq.backend.models.cosmos.HostCosmos;
 import com.noq.backend.repository.cosmos.HostRepositoryCosmos;
 import lombok.RequiredArgsConstructor;
